pyproject.toml
[project]
name = "rdf2vecgpu"
version = "0.2.3"
description = "GPU-Accelerated RDF2Vec – A high-performance GPU implementation of RDF2Vec that uses CUDA and RAPIDS to generate scalable, embeddings for large dense knowledge graphs."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "cudf-cu12>=25.6.0",
    "cugraph-cu12==25.6.*",
    "dask-cudf-cu12>=25.6.0",
    "dask[complete]>=2025.5.0",
    "lightning>=2.5.2",
    "loguru>=0.7.3",
    "numpy>=2.2.6",
    "pydantic>=2.12.3",
    "rdflib>=7.1.4",
    "tables>=3.10.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rdf2vecgpu"]

[tool.hatch.build.targets.sdist]
include = ["src/rdf2vecgpu"]

[project.urls]
"Homepage" = "https://github.com/MartinBoeckling/rdf2vecgpu"
"Issues" = "https://github.com/MartinBoeckling/rdf2vecgpu/issues"

[project.optional-dependencies]
mlflow = [
    "mlflow-skinny>=3.6.0",
]
wandb = [
    "wandb>=0.23.0",
]
test = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["."]
